[TreeModel] Avoid pathchecks on root post_delete
Closes #99

On root Section and Property deletes the old
path is only one item long. Trying to map this
path against an iterator derived from the parent
of the removed object lead to an internal GTK
treemodel error. A new check makes sure that
no root path is mapped against a parent derived
iterator.